Player collision check update!
Changed the way collision with obstacles works. Now the car has a new object with box collider and a script, which stores information for collisions with obstacles. I use this script, to prevent player from moving sideways into obstacles.

Along with that, I added a ray infront of the car and I use it to tell, if car ran into an obstale. If ray intersection is true, then car stops and the game is over.